Update react.mdx
create-react-app has been recently deprecates and is no longer a recommended way to start a React App from scratch:

https://react.dev/blog/2025/02/14/sunsetting-create-react-app
https://react.dev/learn/build-a-react-app-from-scratch

I'm not sure what exactly to replace the CRA command here with, so I simply took the first recommended command from the docs. It uses Vite. In my opinion, of the 3 officially recommended build tools (Vite, Parcel, Rsbuild), Vite is also best suited for creating a sample React app fast.

Let me know if listing commands for all 3 tools is better, or if this change is unnecessary.